requirements.txt
Flask==2.3.2
PyJWT==2.7.0
Werkzeug==2.3.7
bleach==6.0.0
MarkupSafe==2.1.3
pytest==7.4.0
bandit==1.7.5
safety==2.3.5
pbr==5.11.1
python-dotenv==1.0.0